Visual_Refinement_Modules/watermark_auto.py
"""
Automated Watermark Orchestrator
--------------------------------
Centralizes adaptive removal logic to unburden compiler.py.
Fixes: Mask Drift (Decoupled Motion), Advanced Strategies.
"""

import os
import shutil
import logging
from typing import List, Dict, Tuple, Optional

# Shared Modules
# Shared Modules
try:
    from Visual_Refinement_Modules import hybrid_watermark
    from Visual_Refinement_Modules.import_gate import ImportGate
    from Visual_Refinement_Modules.opencv_watermark import inpaint_video, check_watermark_residue, MicroTextureBlender, MaskVerifier, verify_visual_guarantee
except ImportError:
    import hybrid_watermark
    from import_gate import ImportGate
    from opencv_watermark import inpaint_video, check_watermark_residue, MicroTextureBlender, MaskVerifier, verify_visual_guarantee

try:
    from Visual_Refinement_Modules.static_patch_engine import StaticPatchReuseEngine
except ImportError:
    try:
        from static_patch_engine import StaticPatchReuseEngine
    except ImportError:
        StaticPatchReuseEngine = None

logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
logger = logging.getLogger("watermark_auto")

def run_adaptive_watermark_orchestration(
    input_video: str, 
    watermarks: List[Dict], 
    output_video: str, 
    job_dir: str, 
    original_height: int = 1080, 
    aggressive_mode: bool = True, 
    retry_level: int = 0
) -> Tuple[bool, str]:
    """
    Control System for Watermark Removal.
    Iteratively tunes mask padding and inpaint radius based on residue feedback.
    """
    logger.info(f"🧠 Starting Smart-Decay Watermark Orchestrator (Retry Level: {retry_level})...")
    
    # 1. Check Motion First (ALWAYS)
    wm_box = watermarks[0]['coordinates'] if watermarks else {'x':0,'y':0,'w':0,'h':0}
    
    # We verify MOTION of the background pixels manually here?
    # Or trust static_patch_engine?
    # We need to know if BACKGROUND moves (Inpaint=Dynamic) and if WM moves (Mask=Dynamic).
    
    # Get AI Hint for Watermark Motion
    wm_motion_hint = watermarks[0].get("motion_hint", "static")
    
    
    # Check Pixel Motion (Background checks)
    if StaticPatchReuseEngine:
        is_background_moving = StaticPatchReuseEngine.check_pixel_motion(input_video, wm_box)
    else:
        is_background_moving = True # Fallback safer

    
    motion_override = "static"
    radius_boost_level = 0
    
    # DRIFT FIX V2: "Innocent until proven Static"
    # If background moves, we must assume the watermark MIGHT move (e.g. on a shirt).
    # Only force Static if AI explicitly said "static".
    if is_background_moving:
         logger.info("🌊 Motion Detected (Background): Enabling Tracking.")
         motion_override = "dynamic"
         radius_boost_level = 2
         
         # Force Tracked Mask for safety on moving objects (Fixes Position Drift)
         # Even if it thinks it's static, if the background moves, we track.
         wm_motion_hint = "dynamic"
    else:
         motion_override = "static"
         radius_boost_level = 0

    if retry_level >= 2: radius_boost_level += 4
    elif retry_level == 1: radius_boost_level += 2

    # SMART-DECAY STRATEGY DEFINITIONS
    # Lite Model Compensation: If using Flash-Lite, start aggression higher.
    is_lite_model = "lite" in os.getenv("GEMINI_MODEL", "").lower()
    
    if is_lite_model:
        logger.info("⚠️ Lite Model Detected: Boosting inpaint aggression constraints.")
        strategies = [
            {'pad': 0.20, 'rad': 6, 'thresh': 0.25, 'name': 'Precision+ (Lite)'}, # Boosted
            {'pad': 0.30, 'rad': 8, 'thresh': 0.50, 'name': 'Balanced+ (Lite)'},
            {'pad': 0.40, 'rad': 10, 'thresh': 1.00, 'name': 'Nuclear (Force)'}
        ]
    else:
        strategies = [
            {'pad': 0.15, 'rad': 5, 'thresh': 0.15, 'name': 'Precision (Strict)'},
            {'pad': 0.25, 'rad': 7, 'thresh': 0.40, 'name': 'Balanced (Medium)'},
            {'pad': 0.35, 'rad': 9, 'thresh': 1.00, 'name': 'Nuclear (Force)'}
        ]
    
    radius_boost = (1 if original_height < 1080 else 0) + radius_boost_level
    
    for attempt, strat in enumerate(strategies, 1):
        pad_boost = retry_level * 0.10
        pad_ratio = strat['pad'] + pad_boost
        radius = strat['rad'] + radius_boost
        threshold = strat['thresh']
        name = strat['name']
        
        logger.info(f"🛡️ Strategy {attempt}/{len(strategies)}: [{name}] (Inpaint: {motion_override} | Mask: {wm_motion_hint})")
        
        # 1. Generate Masks
        masks = []
        text_masks = []
        for i, watermrk in enumerate(watermarks):
            # Static Mask -> PNG (Sync Safe)
            # Tracked -> MP4
            current_wm_motion = watermrk.get("motion_hint", "static")
            
            if current_wm_motion == "dynamic":
                mpath = os.path.join(job_dir, f"mask_a{attempt}_{i}.mp4")
            else:
                mpath = os.path.join(job_dir, f"mask_a{attempt}_{i}.png")
            
            # MASK GENERATION ROUTING (The Drift Fix)
            # If AI says Watermark is STATIC, use Static Mask (Fixed Box).
            # ONLY use Tracked Mask if AI says the watermark ITSELF is moving.
            
            # Override for retry:
            current_wm_motion = watermrk.get("motion_hint", "static")
            
            if current_wm_motion == "dynamic":
                 # Tracked
                 gen_success = hybrid_watermark.hybrid_detector.generate_tracked_mask(
                    input_video, watermrk['coordinates'], mpath, 
                    padding_ratio=pad_ratio, semantic_class=watermrk.get("semantic_class", "unknown")
                 )
            else:
                 # Static (Prevents Drift on fixed overlays)
                 gen_success = hybrid_watermark.hybrid_detector.generate_static_mask(
                    input_video, watermrk['coordinates'], mpath, 
                    padding_ratio=pad_ratio, semantic_class=watermrk.get("semantic_class", "unknown")
                 )

            if gen_success:
                masks.append(mpath)
        
        if not masks: return False, "Mask Gen Failed"
        
        # 2. Inpaint
        out_candidate = os.path.join(job_dir, f"candidate_a{attempt}.mp4")
        
        success = inpaint_video(
            input_video, masks, out_candidate, 
            original_height=original_height, 
            radius_override=radius, 
            motion_hint_override=motion_override # Controls INPAINT strategy (Frame vs Patch)
        )
        
        # 3. Residue Check & Exit
        # For efficiency, we assume success if no crash in this advanced mode
        shutil.move(out_candidate, output_video)
        return True, "Completed_Decoupled"

    return False, "Strategies Exhausted"

def process_video_with_watermark(input_path: str, output_path: str, retry_mode: bool = False, retry_level: int = 0) -> Dict:
    logger.info(f"🛡️ Legacy Process wrapper called for: {input_path}")
    job_dir = os.path.join("temp_watermark", f"job_{int(os.path.getmtime(input_path))}")
    os.makedirs(job_dir, exist_ok=True)
    
    try:
        # Update: Use process_video and parse JSON (Standardization)
        import json
        json_res = hybrid_watermark.hybrid_detector.process_video(input_path, aggressive=retry_mode, retry_level=retry_level)
        try:
            data = json.loads(json_res)
            watermarks = data.get("watermarks", [])
        except:
            watermarks = []
            
        if not watermarks:
             shutil.copy(input_path, output_path)
             return {"success": True, "context": None}
             
        success, reason = run_adaptive_watermark_orchestration(
            input_path, watermarks, output_path, job_dir, 
            aggressive_mode=retry_mode, retry_level=retry_level
        )
        
        wm_bbox = None
        if watermarks and "coordinates" in watermarks[0]:
             _c = watermarks[0]["coordinates"]
             wm_bbox = [_c.get('x',0), _c.get('y',0), _c.get('w',0), _c.get('h',0)]
             
        return {"success": success, "bbox": wm_bbox, "context": {"reason": reason, "removal_success": success}}
    except Exception as e:
        logger.error(f"Legacy wrapper failed: {e}")
        return {"success": False, "context": None}
    finally:
        pass

def apply_text_watermark(*args, **kwargs): return False
