Visual_Refinement_Modules/watermark_requirements.txt
# Watermark Modules Dependencies
opencv-python
numpy
python-dotenv
google-generativeai
gradio
python-telegram-bot
Pillow
onnxruntime
